+++ client/src/contexts/bible.tsx
@@ -19,6 +19,9 @@ type IBibleContext = {
     nextVerse: () => void;
     toggleSelectedVerse: (verse: IVerse) => void;
     clearSelectedVerses: () => void;
+    copySelectedVerses: () => void;
+    shareSelectedVerses: () => void;
+    formatSelectedVerses: () => string;
     applyHighlightColor: (color: string) => void;
 };
 
@@ -88,6 +91,65 @@ const BibleProvider = ({ children }: { children: React.ReactNode }) => {
 
     const clearSelectedVerses = () => { setSelectedVerses([]); }
 
+    const copySelectedVerses = () => {
+        const verses = selectedVerses.sort((a, b) => a - b);
+
+        const versesText = verses.map((v) => {
+            const verse = book?.verses?.find((verse) => verse.number === v);
+            return verse ? `${verse.number} ${verse.text}` : "";
+        });
+
+        const versesTextJoined = versesText.join("\n") + "\n" + `${book?.name} ${formatSelectedVerses()} (${version?.toUpperCase()})`;
+
+        navigator.clipboard.writeText(versesTextJoined);
+
+        toast({ title: "Versículos copiados!", duration: 800, });
+
+        clearSelectedVerses();
+    };
+
+    const shareSelectedVerses = () => {
+        const textSelected = book?.verses?.filter((verse) =>
+            verse.number && selectedVerses.includes(verse.number)
+        );
+    
+        let textFormatted = '';
+    
+        textSelected?.forEach((verse) => {
+            textFormatted += `${verse.number} ${verse.text} `;
+        });
+    
+        navigate('/share', { state: { text: textFormatted, book: {name: book?.name, chapter, version, format: formatSelectedVerses()} } });
+    };
+    
+
+    const formatSelectedVerses = (): string => {
+        if (selectedVerses.length === 0)
+            return '';
+
+        const numbers = [...selectedVerses].map(verse => Number(verse)).sort((a, b) => a - b);
+
+        let numbersArray: number[][] = [];
+        let referenceIndex = 0;
+
+        numbers.forEach((number, index) => {
+            if (index === numbers.length - 1 || numbers[index + 1] - number > 1) {
+                const start = numbers[referenceIndex];
+                const end = number;
+
+                if (start !== end) {
+                    numbersArray.push([start, end]);
+                } else {
+                    numbersArray.push([start]);
+                }
+
+                referenceIndex = index + 1;
+            }
+        });
+
+        return numbersArray.map(arr => arr.join('-')).join(', ');
+    };
+
     const prevVerse = () => {
         if (chapter === 1) {
             if (book?.etc.prevBook) {
@@ -150,9 +212,9 @@ const BibleProvider = ({ children }: { children: React.ReactNode }) => {
         }
     };
 
-
     const contextValue = useMemo(
-        () => ({
+        () => { 
+            return ({
             version: version ?? "nvi",
             book: {
                 abbrev: book?.abbrev ?? "gn",
@@ -174,8 +236,11 @@ const BibleProvider = ({ children }: { children: React.ReactNode }) => {
             nextVerse,
             toggleSelectedVerse,
             clearSelectedVerses,
+            copySelectedVerses,
+            shareSelectedVerses,
+            formatSelectedVerses,
             applyHighlightColor,
-        }),
+        })},
         [version, book, chapter, isLoading, error, selectedVerses, highlightedVerses]
     );